[ALS-5535] Implement route restriction and remove API display function (#22)
* [ALS-5535] Implement route restriction and remove API display function

Added a list of allowed routes in the frontend application and implemented a check to ensure only valid routes are accessed. If invalid, the user is redirected to the landing page. The function for displaying API has also been removed from router.js.

* [ALS-5535] Update route validation in router.js

Modified the route validation logic in router.js. The code now checks for the inclusion of callback.name.split('/')[1] in the allowedRoutes, instead of just the 'name' parameter. This is done to ensure accurate route identification and permission checking.