Implement automatic reauthentication on session expiry
A middleware has been added for automatic user reauthentication when the session is near expiration or expired. Additionally, behavior for unauthorized responses is updated to clear session storage and reload the page instead of being redirected to "not authorized" page. These changes provide smoother user experience and improved session management.