Add footer.js for overriding title and redirect handling
The footer.js file overrides the render function, allowing for altered page title handling based on the path segment. It also applies reauthentication middleware and handles redirects for links that open in a new tab. The changes aim to provide better functionality and improved user experience for the PIC-SURE's BioData Catalyst.