Improve reauthentication middleware functionality
A conditional has been added to the ajaxPrefilter function in the reauthenticate.js file to skip reauthentication for urls starting with '/psamaui'. Also, the 'session.init' function call has been replaced with 'session.sessionInit' to adhere to the updated method name. These modifications aim to improve token expiration management and user reauthentication processes in the application.