auto re-authentication fix (#35)
* Implement automatic reauthentication on session expiry

A middleware has been added for automatic user reauthentication when the session is near expiration or expired. Additionally, behavior for unauthorized responses is updated to clear session storage and reload the page instead of being redirected to "not authorized" page. These changes provide smoother user experience and improved session management.

* Add footer.js for overriding title and redirect handling

The footer.js file overrides the render function, allowing for altered page title handling based on the path segment. It also applies reauthentication middleware and handles redirects for links that open in a new tab. The changes aim to provide better functionality and improved user experience for the PIC-SURE's BioData Catalyst.

* Improve reauthentication middleware functionality

A conditional has been added to the ajaxPrefilter function in the reauthenticate.js file to skip reauthentication for urls starting with '/psamaui'. Also, the 'session.init' function call has been replaced with 'session.sessionInit' to adhere to the updated method name. These modifications aim to improve token expiration management and user reauthentication processes in the application.

* Update session management and reauthentication

Revamped the reauthentication.js middleware in open-pic-sure-bdc-frontend. The updated code adds a more robust token handling mechanism, with improved handling of token refresh scenarios, a scheduled check on token expiration status, and a conditional check to skip reauthentication for specific URLs. The recent changes aim to significantly improve the user experience regarding session management and re-authentication.